This is our original translation Rcode file
We used the "Wix" SaaS to generate a website that for sharing class resources publicly, where the whole website is not included in the Github since not all parts of our website were created by our coding. This is the second part of our website, which is a specific topic .Rmd tutorial (we are translating the external resources on Mosaic plot topic). 

Since we didn't write the code for the Wix website excluding the 2nd part of the website that we wrote with .Rmd, we only include the 2nd part in the Github.